ENH: Support almost unlimited summary preprocessing.
Add a configuration option to reload SUMMARY.md, evaluate it as a template, and
re-parse chapters. This allows us to support full arbitrary preprocessing of
summary.md, but it means we discard the results of any preprocessors that ran
before us.

In most scenarios this will probably be an OK tradeoff. The ideal solution
still requires something like https://github.com/rust-lang/mdBook/issues/2466.